refactor: IssueListPage에 속한 컴포넌트 구조 변경
- webpack config 파일에 jsx extension 추가
- App.js에 있는 컴포넌트들을 개발 컴포넌트파일로 변경
