+++ frontend/src/App.js
@@ -1,74 +1,6 @@
 import React from 'react';
 
-
-const IssueListNav = () => {
-  return (
-    <div>
-      이슈리스트메뉴
-    </div>
-  )
-}
-
-const SearchInput = () => {
-  return (
-    <div>
-      searchInput
-    </div>
-  )
-}
-
-const PropertyButton = () => {
-  return (
-    <div>
-      propertybutton
-    </div>
-  )
-}
-
-const NewIssueButton = () => {
-  return (
-    <div>
-      newIssueButton
-    </div>
-  )
-}
-
-const SearchBar = () => {
-  return (
-    <div>
-      <SearchInput/>
-      <PropertyButton/>
-      <NewIssueButton/>
-    </div>
-  )
-}
-
-const IssueListItem = () => {
-  return (
-    <div>
-      Issue1
-    </div>
-  )
-}
-
-const IssueList = () => {
-  return (
-    <div>
-      <IssueListItem />
-    </div>
-  )
-}
-
-const IssueListPage = () => {
-  return (
-    <div>
-      <SearchBar />
-      <IssueListNav />
-      <IssueList />
-    </div>
-  );
-}
-
+import IssueListPage from './pages/IssueListPage'
 
 const App = () => {
   return (
