+++ frontend/src/components/IssueList.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import IssueListItem from './IssueListItem';
+
+const IssueList = () => {
+  return (
+    <div>
+      <IssueListItem />
+    </div>
+  )
+}
+
+export default IssueList;
