+++ frontend/src/components/IssueListItem.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const IssueListItem = () => {
+  return (
+    <div>
+      Issue1
+    </div>
+  )
+}
+
+export default IssueListItem;
