+++ frontend/src/components/IssueListNav.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const IssueListNav = () => {
+  return (
+    <div>
+      이슈리스트메뉴
+    </div>
+  )
+}
+
+export default IssueListNav;
