+++ frontend/src/components/NewIssueButton.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const NewIssueButton = () => {
+  return (
+    <div>
+      newIssueButton
+    </div>
+  )
+}
+
+export default NewIssueButton;
