+++ frontend/src/components/PropertyButton.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const PropertyButton = () => {
+  return (
+    <div>
+      propertybutton
+    </div>
+  )
+}
+
+export default PropertyButton;
