+++ frontend/src/components/SearchInput.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const SearchInput = () => {
+  return (
+    <div>
+      searchInput
+    </div>
+  )
+}
+
+export default SearchInput;
