+++ frontend/src/pages/IssueListPage.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import SearchBar from '../components/SearchBar';
+import IssueListNav from '../components/IssueListNav';
+import IssueList from '../components/IssueList';
+
+const IssueListPage = () => {
+  return (
+    <div>
+      <SearchBar />
+      <IssueListNav />
+      <IssueList />
+    </div>
+  );
+}
+
+export default IssueListPage;
