+++ frontend/webpack.config.js
@@ -33,4 +33,7 @@ module.exports = {
     new HtmlWebpackPlugin({
       template: './src/index.html',
     })],
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  }
 };
